--- client/src/components/Dashboard/Reviews/Review.jsx
@@ -1,16 +1,0 @@
-
-// const Reseña = () => {
-//   return <div className={style.userCard} key={index}>
-//         <img className={style.userIcon} src="https://i0.wp.com/lamiradafotografia.es/wp-content/uploads/2014/07/foto-perfil-psicologo-180x180.jpg?resize=180%2C180"></img>
-//         <div className={style.nameAndUser}>
-//             <p className={style.name}>{u.local}</p>
-//             <p className={style.usernames}>{u.name}</p>
-//         </div>
-//         <div className={style.titleAndStars}>
-//             <p className={style.name}>{u.title}</p>
-//             <p className={style.usernames}>{u.stars} estrellas</p>
-//         </div>
-//         <button className={style.buttonExaminar} onClick={() => fn(3)}>Examinar</button>
-//     </div>;
-// };
-// export default Reseña;
